pyproject.toml
[project]
name = "richard-x46-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=6.0.0",
    "branca>=0.8.2",
    "folium>=0.20.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "streamlit>=1.52.0",
    "streamlit-folium>=0.25.3",
]
